Add PLAY_CALLER: and PLAY_BOTH: to MACRO_RESULT responses
To allow playing notification messages for checking whether or not both sides
of a call are secure (as defined by the user in the dialplan), you can play
a message by setting MACRO_RESULT to either PLAY_CALLER:prompt or
PLAY_BOTH:prompt_caller^prompt_callee with the prompt options being the name
(without extension) of a prompt.  PLAY_BOTH: will play to both sides of the
call simultaneously just before bridging the call.