Make sure ANSWERED_ELSEWHERE works in combination with chan_local
In order for this to work, I had to check the calling channels flag and force
app_dial and app_queue to set it on the new (outbound) channel regardless
of the option to queue and dial. Now, queue members using local/123@context
as members will not get missed calls if queue is called with the C option.