Fix another race condition in app_confbridge
If a marked user entered the conference while an unmarked
user was listening to the prompt telling him that he is
waiting for the leader, then badness would happen. I'll
not elaborate further. It's fixed now though.

Because I'm such a good guy, I also got rid of
extra whitespace in this commit, too.