Creating a branch to create a method of automatic channel fixup during a masquerade
Currently, there are a few structures in Asterisk which have "fixup" callbacks
that are run during masquerades to ensure that their channel pointers point to the
correct structure. While having the fixup callback available is nice, we keep coming
to more and more places where we have to use datastores for the sole purpose of
making sure that a channel pointer is correct.

The idea behind this branch is to make a generic method of specifying a channel
pointer in your code in a way that it will automatically update when a masquerade
occurs.

I got the idea to make this branch when I saw issue #14374 in the bugtracker.
The idea is to genericize this method of correcting channel pointers instead of having
to re-implement the same datastore over and over. I will essentially be using what's
implemented in app_chanspy and copying its logic to a more generic place so that the
logic may be used in app_chanspy as well as app_mixmonitor. I will also do an audit
of the code to see if there are other areas prone to the same sort of problems.