Initial commit of new "autochan" work
What is an autochan? Well, I'm sure glad you asked! It's a means
by which one may have a pointer to a channel which is "owned" by
a separate thread from your own and which will automatically update
should the channel that you were pointing to be masqueraded.

This commit introduces the basic API as well as changes to
app_mixmonitor to make use of it.

The following still needs to be done:

1. Add doxygen to the autochan API
2. Convert app_chanspy to use an autochan
3. Find other areas in the code which may benefit from
   the use of an autochan