a bunch of build system changes, designed to actually make the Asterisk symbol namespace pollution problems actually be solved. the strip_nonapi script was an attempt to fix this, but failed completely because it was modifying the wrong symbol table.
with these changes, only the symbols that *should* be exported from the main asterisk binary, and res_* modules, are actually exported, everything else is hidden. this is true even for modules that link to external static libraries, where previously all the symbols from those libraries would be exported into the global symbol namespace as well.