Remove the NOTNOW sections from pbx.c
These sections appeared to override the hashtab lookups
and instead do a traversal of the list of contexts or extensions.
Unfortunately the logic used here was not always correct, and 
I can think of no reason that you would want to use the slower
lookup. This may have been useful at some point, but now it is
not and just adds confusion to the code.