Tweak the autofill and ringall behavior in num_available_members
This change makes it so that we don't always report that there is a member
available if there isn't really one available. Still, we will break out of the
loop as soon as we realize there is at least one member available since that's
really all we care about in this particular case.