I'm abandoning this branch for multiple reasons: I changed commenting strategies twice in the middle. I found that I may have unneccessarily been making changes that were not needed.
I'm hoping with the recent increased effort pertaining to Doxygen documentation
that things will evolve to a more compliant state.

The biggest observed problem was simply not documenting all of the parameters.
The other common problem was not properly escaping Doxygen reserved commands or 
symbols.

The unnecessary changes made are related to something I said in #asterisk-dev
a couple of weeks ago. I said that if Doxygen documentation existed for both 
the function definition and declaration, that the comments were improperly
merged ignoring those with the declaration. I began to question this behavior
and can't seem to replicate that now. Perhaps the rules for merging are
dependent on which tags are being used.