Fix crash that could occur on SIP attended transfers of unbridged calls.
If trying to transfer a Meetme conference or an IVR where there is no bridged
channel that is being transferred, there is no collected line information
to gather. Trying to do so means accessing NULL pointers and generally leads
to horrible suffering, anguish, and other hyperbole to describe a crash.

AST-200