This I believe fixes the channel mappings for NFAS.
The problem was that the channel number being sent to libpri was incorrectly 
calculated based on the physical span instead of its mapping. The channel
number must be offset by the number of b-chans already present in the same
trunkgroup. For clarity, the channel listing in astdb now contains the physical
span in addition to the channel number.