Create a branch for addressing issue #13630.
The issue has to do with the fact that spiral support in chan_sip was implemented
pretty badly. The worst part probably had to do with the fact that you could only
spiral a request to another SIP endpoint.

The fix that has been added to 1.4 right now is to treat a SIP spiral as a call forward.
This fix makes good sense in my mind, since it means that we will be taking the proper
steps to accommodate the SIP spiral. The problem right now is that the only methods that
recognize a call_forward currently are app_dial, app_queue, and the ast_dial API. This
means that using a call file, an originate command, or a transfer will not allow for
recognition of a SIP spiral.

Most of the actual spiral handling in chan_sip is taken care of in 1.4 at this point, it's
just a matter of fixing the areas where call forwards are not recognized. The biggest
portion of this work will be creating a generic ast_forward_call function in channel.c that
is modeled after the call-forwarding logic in app_dial. This will be used to handle call forwards
from outside of the areas that already support call forwarding.