Creating a branch to handle digit manipulation of connected line updates.
Currently, when someone executes an application like Dial or Queue, they may receive
a connected line update. However, since control has left the dialplan, they are unable
to make any changes to that information when passing it along to the other side. Business
logic may dictate that certain prefixes are placed on numbers or names are edited in
some way depending on the destination of the connected line update.

The plan for this branch is to add some user-defined macros which may be called at different
points during a call so that control may return to the dialplan. Inside these macros, the
connected line data may be manipulated to allow for the aforementioned business logic to
be applied.